fogl-102 Rename starter to controller
