--- src/python/foglamp-config.example.yaml
@@ -1,17 +1,0 @@
-####################
-# The path to this file can be specified by the FOGLAMP_CONFIG_PATH
-# environment variable
-####################
-
-####################
-# Database
-####################
-
-database:
-  host: localhost
-  port: 5432
-  user: postgres
-  password: postgres # Can be overridden by the FOGLAMP_DB_PASSWORD environment variable
-  db: foglamp
-  driver: postgresql
-
