+++ src/python/foglamp/admin_api/wsgi.py
@@ -9,7 +9,7 @@ gunicorn foglamp.admin_api.wsgi:app --bind localhost:8080 --worker-class aiohttp
 """
 
 import logging
-from .app import build as build_app
+from .app_builder import build as build_app
 
 
 try:
