+++ src/python/foglamp/model/controller.py
@@ -1,4 +1,6 @@
-from .config import read
+from .env import read
+
 
 def start():
     read()
+
